// src/common/enums/reservationStatusEnum.ts
export enum ReservationStatusEnum {
  PENDING = "PENDING",
  APPROVED = "APPROVED",
  REJECTED = "REJECTED",
  CANCELLED = "CANCELLED",
  CANCELLED_BY_GUEST = "CANCELLED_BY_GUEST",
  CANCELLED_BY_HOST = "CANCELLED_BY_HOST",
  CHECKED_IN = "CHECKED_IN",
  NO_SHOW = "NO_SHOW",
  CHECKED_OUT = "CHECKED_OUT",
}